pyproject.toml
[project]
name = "ares"
version = "0.1.0"
description = "ARES - Automated Research Engine for Specification"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "openai>=1.68.0",
  "pypdf>=5.4.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "rich>=13.9.0",
  "tenacity>=8.2.3",
  "ulid-py>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "core"]
